+++ bot/utils/insta_dl_utils.py
@@ -1,3 +1,4 @@
+import json
 import math
 import os
 import secrets
@@ -8,7 +9,7 @@ from typing import Optional
 
 from bot.config import bot, conf
 from bot.fun.emojis import enhearts
-from bot.pkg.insta_dl import download_instagram
+from bot.pkgs.insta_dl import download_instagram
 
 from .bot_utils import (
     hbs,
@@ -74,6 +75,10 @@ class InstagramHelper:
     @property
     def name(self):
         return self._listener.name
+
+    @property
+    def download_is_complete(self):
+        return self._listener.completed
 
     async def _on_download_progress(self, current: int, total: int, file_path: str):
         """Called after each chunk; updates internal state and edits the message."""
